* origin/topic/vlad/gh-1225:
  Extract length as a uint8
  Change ICMP ND length to a uint16

(cherry picked from commit 5e164469f34ddd894c59eb6304707766f8ae6d4a)

+++ src/analyzer/protocol/icmp/ICMP.cc
@@ -745,7 +745,7 @@ VectorVal* ICMP_Analyzer::BuildNDOptionsVal(int caplen, const u_char* data)
 			}
 
 		uint8 type = *((const uint8*)data);
-		uint8 length = *((const uint8*)(data + 1));
+		uint16 length = *((const uint8*)(data + 1));
 
 		if ( length == 0 )
 			{
